Update settings.md
We are no longer allowing folks to register for NPR Story API keys as this API is slated for deprecation. So we are sending them to the DS ticketing system to ask for a key. If they are the general public they are encouraged to contact Audience Relations at the link added here.